refactor: 重构截图功能
